Separate out model generation code
