+++ Makefile
@@ -7,7 +7,7 @@ $(eval $(call golang-version-check,1.6))
 
 test:
 	rm -rf generated/*
-	go run main.go genclients.go -file swagger.yml -package $(PKG)/generated
+	go run main.go genclients.go genmodels.go -file swagger.yml -package $(PKG)/generated
 	cd impl && go build
 	# Temporarily run the client here since it isn't used in impl
 	cd generated/client && go build
